refactor(deposit): move song-block UI orchestration into DepositSong
